Use pkg/fulcioroots and pkg/tuf from sigstore/sigstore
